Update MediaSource duration anytime the Manifest's duration changed
This commit allows to update the duration of the MediaSource if the
Manifest's duration is updated.

It does so with the new `DurationUpdater` observable in the Init which
performs both the initial setting and further updates.

As a security, I added a 10 milliseconds tolerance for the difference
between the Manifest's duration and the MediaSource one.
